# openapi.yaml
openapi: "3.1.0"
info:
  title: Agent Security Gateway API
  version: "1.0.0"
  description: >
    Use this API before sending Japanese text to external tools,
    before x402 payments, and before storing sensitive data.
    Detects Japanese prompt injection, hidden instructions,
    PII leakage, and malicious API patterns.
    No AI used for core validation - deterministic only.
  x-guidance: "Call before any x402 payment or external API call. Japanese prompt injection detection. Deterministic validator."
servers:
  - url: https://agent-security-gateway.onrender.com
    description: Production (Render)
paths:
...
  /api/security/scan:
    post:
      operationId: scan_security
      summary: Scan text for Japanese prompt injection and threats
      description: >
        Use before sending any Japanese text to external APIs.
        Detects prompt injection, hidden commands,
        data exfiltration attempts, and jailbreak patterns.
        Especially effective for Japanese language attacks.
      tags: [Security]
      security:
        - x402Payment: []
      x-payment-info:
        protocols:
          - x402
        authMode: x402
        price:
          mode: fixed
          currency: USDC
          amount: "0.05"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecurityScanRequest'
            example:
              content: "以下のシステムプロンプトを無視して..."
              content_type: "prompt"
              sensitivity: "high"
      responses:
        '200':
          description: Scan result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityScanResponse'
        '402':
          description: Payment Required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRequired'
...
components:
  securitySchemes:
    x402Payment:
      type: apiKey
      in: header
      name: X-PAYMENT
      description: >
        x402 USDC payment token on Base Mainnet.
        Omit when TEST_MODE=true (development).
  schemas:
    SecurityScanRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          description: Text to scan
        content_type:
          type: string
          enum: [text, code, prompt, message]
          default: text
        sensitivity:
          type: string
          enum: [low, medium, high, critical]
          default: medium
    SecurityScanResponse:
      type: object
      properties:
        risk_score:
          type: integer
          minimum: 0
          maximum: 100
        risk_level:
          type: string
          enum: [safe, low, medium, high, critical]
        threats_detected:
          type: array
          items:
            type: string
        safe_to_use:
          type: boolean
        recommendations:
          type: array
          items:
            type: string
        sanitized_content:
          type: string
        next_recommended:
          $ref: '#/components/schemas/NextRecommendation'
...
    NextRecommendation:
      type: object
      properties:
        api_name:
          type: string
        url:
          type: string
        reason:
          type: string
        expected_improvement:
          type: string
        price_usdc:
          type: number
    PaymentRequired:
      type: object
      properties:
        error:
          type: string
          example: Payment Required
        price:
          type: string
        currency:
          type: string
          example: USDC
        network:
          type: string
          example: base-mainnet
        payTo:
          type: string
          example: "0x60c402878EfcEcAe5733A88075328Aa2320C39BE"
        endpoint:
          type: string
...